requirements.txt
streamlit
pandas
scikit-learn
matplotlib
seaborn
joblib
fpdf2
numpy
tensorflow
xgboost
